backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-dateutil==2.8.2
spacy==3.7.2
icalendar==5.0.11
pytz==2023.3
